// src/libs/api/schema/accounting.d.ts
export interface components {
    schemas: {
        /** Account */
        Account: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            /** Number */
            number: string;
            account_type: components["schemas"]["AccountType"] | null;
            /** Balance */
            balance: string;
            /** Account Number */
            account_number: string;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Last Modified At
             * Format: date-time
             */
            last_modified_at: string;
            created_by: components["schemas"]["UserMin"];
            last_modified_by: components["schemas"]["UserMin"] | null;
            parent: components["schemas"]["AccountMin"] | null;
            currency: components["schemas"]["CurrencyMin"];
            /** Level */
            level: number;
        };
        /** AccountBalanceUpdate */
        AccountBalanceUpdate: {
            /** Balance */
            balance: number | string;
        };
        /** AccountCreate */
        AccountCreate: {
            /** Name */
            name: string;
            /** Number */
            number: string;
            account_type: components["schemas"]["AccountType"] | null;
            /** Parent Id */
            parent_id: number | null;
            /** Currency Id */
            currency_id: number;
        };
        /** AccountMin */
        AccountMin: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            /** Number */
            number: string;
            account_type: components["schemas"]["AccountType"] | null;
            /** Account Number */
            account_number: string;
        };
        /** AccountOpt */
        AccountOpt: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            /** Number */
            number: string;
            account_type: components["schemas"]["AccountType"] | null;
            /** Account Number */
            account_number: string;
        };
        /**
         * AccountType
         * @enum {string}
         */
        AccountType: "Asset" | "Liability" | "Equity" | "Revenue" | "Expense";
        /** AccountUpdate */
        AccountUpdate: {
            /** Name */
            name: string;
            /** Number */
            number: string;
            account_type: components["schemas"]["AccountType"] | null;
            /** Parent Id */
            parent_id: number | null;
            /** Currency Id */
            currency_id: number;
        };
        /** CurrencyMin */
        CurrencyMin: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            /** Iso */
            iso: string;
            /** Symbol */
            symbol: string | null;
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /** PaginatedAccount */
        PaginatedAccount: {
            /** Total Items */
            total_items: number;
            /** Total Page */
            total_page: number;
            /** Page */
            page: number;
            /** Items */
            items: components["schemas"]["Account"][];
        };
        /** PaginatedAccountOpt */
        PaginatedAccountOpt: {
            /** Total Items */
            total_items: number;
            /** Total Page */
            total_page: number;
            /** Page */
            page: number;
            /** Items */
            items: components["schemas"]["AccountOpt"][];
        };
        /** UserImageURL */
        UserImageURL: {
            /** Url */
            url: string | null;
        };
        /** UserMin */
        UserMin: {
            /** Pk */
            pk: number;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /**
             * Email
             * Format: email
             */
            email: string;
            image: components["schemas"]["UserImageURL"];
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}